% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{nclRpackage-package}
\alias{nclRpackage}
\alias{nclRpackage-package}
\alias{nclrpackage}
\title{R courses at Newcastle University}
\description{
Functions and datasets used in for R
courses at Newcastle University. This package is used in
the "Building a package" course".
}
\keyword{package}

